@@ -10,26 +10,31 @@
     "test": "NODE_ENV=test jest"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
+    "@emotion/cache": "^11.10.1",
+    "@emotion/react": "^11.10.0",
+    "@emotion/server": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@mui/material": "^5.10.0",
+    "@mui/system": "^5.10.0",
     "classnames": "^2.3.1",
-    "color": "^4.2.3",
     "copy-to-clipboard": "^3.3.1",
     "next": "12.1.6",
     "next-translate": "^1.4.0",
     "ramda": "^0.28.0",
-    "react": "18.1.0",
-    "react-dom": "18.1.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@emotion/jest": "^11.10.0",
     "@svgr/webpack": "^6.2.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^14.2.0",
-    "@types/color": "^3.0.3",
+    "@types/jest": "^28.1.6",
     "@types/node": "17.0.31",
     "@types/ramda": "^0.28.14",
-    "@types/react": "18.0.8",
-    "@types/react-dom": "18.0.3",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
     "@types/react-test-renderer": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
     "@typescript-eslint/parser": "^5.22.0",
@@ -47,6 +52,6 @@
     "jest-environment-jsdom": "^28.1.0",
     "react-test-renderer": "^18.1.0",
     "ts-jest": "^28.0.2",
-    "typescript": "4.6.4"
+    "typescript": "^4.7.4"
   }
 }
