@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
-import { ThemeProvider } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
+import { ThemeProvider } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
 import { darkTheme } from '@src/styles/theme';
 import { useApp } from './hooks';
 
