@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { ClassAttributes, HTMLAttributes } from 'react';
 import { render, screen } from '@testing-library/react';
 import Component from '.';
 
 jest.mock('..', () => ({
-  Nav: (props: JSX.IntrinsicAttributes & React.ClassAttributes<HTMLDivElement> & React.HTMLAttributes<HTMLDivElement>) => <div data-testid="Nav" {...props} />,
-  Footer: (props: JSX.IntrinsicAttributes & React.ClassAttributes<HTMLDivElement> & React.HTMLAttributes<HTMLDivElement>) => <div data-testid="Footer" {...props} />,
+  Nav: (props: JSX.IntrinsicAttributes & ClassAttributes<HTMLElement> & HTMLAttributes<HTMLElement>) => <div data-testid="Nav" {...props} />,
+  Footer: (props: JSX.IntrinsicAttributes & ClassAttributes<HTMLElement> & HTMLAttributes<HTMLElement>) => <div data-testid="Footer" {...props} />,
 }));
 // ==================================
 // unit tests
