@@ -1,33 +1,22 @@
-import React from 'react';
+import { FC, HTMLAttributes } from 'react';
 import BDLogo from '@assets/big-dipper-red.svg';
-import { SectionLimit } from '@components';
-import { useStyles } from './styles';
-import HamburgerIcon from './components/hamburger_icon';
-import DropdownMenu from './components/dropdown_menu';
+import { StyledSectionLimit } from './styles';
+import HamburgerIcon from './components/hamburger-icon';
+import DropdownMenu from './components/dropdown-menu';
 import { useNav } from './hooks';
 
-const Nav: React.FC<ComponentDefault> = () => {
-  const classes = useStyles();
-  const {
-    isIcon,
-    isMenu,
-    toggleHamburgerMenu,
-  } = useNav();
+const Nav: FC<HTMLAttributes<HTMLElement>> = () => {
+  const { isIcon, isMenu, toggleHamburgerMenu } = useNav();
 
   return (
     <div>
-      <SectionLimit className={classes.root}>
+      <StyledSectionLimit>
         <div className="wrapper">
-          <BDLogo className="logo" />
-          <HamburgerIcon
-            toggleHamburgerMenu={toggleHamburgerMenu}
-            isIcon={isIcon}
-          />
+          <BDLogo className="logo" height={36} />
+          <HamburgerIcon toggleHamburgerMenu={toggleHamburgerMenu} isIcon={isIcon} />
         </div>
-        <DropdownMenu
-          isMenu={isMenu}
-        />
-      </SectionLimit>
+        <DropdownMenu isMenu={isMenu} />
+      </StyledSectionLimit>
     </div>
   );
 };
