@@ -1,14 +1,23 @@
-import React from 'react';
+import { Children } from 'react';
 import Document, {
   Html, Main, NextScript, Head,
 } from 'next/document';
-import { ServerStyleSheets } from '@material-ui/core/styles';
+import createEmotionServer from '@emotion/server/create-instance';
+import createEmotionCache from '../utils/create-emotion-cache';
 
 export default class MyDocument extends Document {
   render() {
     return (
       <Html lang="en">
-        <Head />
+        <Head>
+          <link
+            rel="preload"
+            href="/fonts/SF-Pro-Text-Regular.woff2"
+            as="font"
+            type="font/woff2"
+            crossOrigin="anonymous"
+          />
+        </Head>
         <title>Big Dipper</title>
         <body>
           <Main />
@@ -23,20 +32,31 @@ export default class MyDocument extends Document {
 // it's compatible with server-side generation (SSG).
 MyDocument.getInitialProps = async (ctx) => {
   // Render app and page and get the context of the page with collected side effects.
-  const sheets = new ServerStyleSheets();
   const originalRenderPage = ctx.renderPage;
+  const cache = createEmotionCache();
+  const { extractCriticalToChunks } = createEmotionServer(cache);
 
   ctx.renderPage = () => originalRenderPage({
-    enhanceApp: (App) => (props) => sheets.collect(<App
-      {...props}
-    />),
+    enhanceApp: (App) => function EnhanceApp(props) {
+      const propsWithCache = { emotionCache: cache, ...props } as unknown as typeof props;
+      return <App {...propsWithCache} />;
+    },
   });
 
   const initialProps = await Document.getInitialProps(ctx);
+  const emotionStyles = extractCriticalToChunks(initialProps.html);
+  const emotionStyleTags = emotionStyles.styles.map((style) => (
+    <style
+      data-emotion={`${style.key} ${style.ids.join(' ')}`}
+      key={style.key}
+      // eslint-disable-next-line react/no-danger
+      dangerouslySetInnerHTML={{ __html: style.css }}
+    />
+  ));
 
   return {
     ...initialProps,
     // Styles fragment is rendered after the app and page rendering finish.
-    styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()],
+    styles: [...Children.toArray(initialProps.styles), ...emotionStyleTags],
   };
 };
